requirements.txt
# Core dependencies
celery>=5.3.0
redis>=4.5.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database and data handling
pandas>=2.0.0
alembic>=1.12.0  # For database migrations

# HTTP and API
requests>=2.31.0
urllib3>=2.0.0

# Testing
pytest>=7.4.0
pytest-celery>=0.0.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Development tools (optional)
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Code linting
mypy>=1.5.0   # Type checking